backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-multipart==0.0.9
azure-ai-projects>=2.0.0b1
azure-identity
github-copilot-sdk
